+++ changelog.md
@@ -2,6 +2,7 @@
 
 ### *2015.06.xx* - 1.5.0
 * **New** - rtl support.
+* **New** - Integration with WooCommerce Deposits.
 
 ### *2015.06.10* - 1.4.6
 * **Fix** - `font-family` delcaration on `select`s.